Updated code docs, small refactors to code
