+++ src/bunit.core/Rendering/RenderTreeFrameCollection.cs
@@ -4,48 +4,30 @@ using Microsoft.AspNetCore.Components.RenderTree;
 
 namespace Bunit.Rendering
 {
+	/// <summary>
+	/// Represents a collection of <see cref="ArrayRange{RenderTreeFrame}"/>.
+	/// </summary>
 	public sealed class RenderTreeFrameCollection
 	{
-		private Dictionary<int, ArrayRange<RenderTreeFrame>> _currentRenderTree = new Dictionary<int, ArrayRange<RenderTreeFrame>>();
+		private readonly Dictionary<int, ArrayRange<RenderTreeFrame>> _currentRenderTree = new Dictionary<int, ArrayRange<RenderTreeFrame>>();
 
+		/// <summary>
+		/// Gets the <see cref="ArrayRange{RenderTreeFrame}"/> associated with the <paramref name="componentId"/>.
+		/// </summary>
+		/// <param name="componentId"></param>
+		/// <returns></returns>
 		public ArrayRange<RenderTreeFrame> this[int componentId] => _currentRenderTree[componentId];
 
+		/// <summary>
+		/// Creates an instance of the <see cref="RenderTreeFrameCollection"/>,
+		/// </summary>
 		internal RenderTreeFrameCollection() { }
 
+		/// <summary>
+		/// Checks whether the collection contains a <see cref="ArrayRange{RenderTreeFrame}"/> for the <paramref name="componentId"/>.
+		/// </summary>
 		public bool Contains(int componentId) => _currentRenderTree.ContainsKey(componentId);
 
 		internal void Add(int componentId, ArrayRange<RenderTreeFrame> frames) => _currentRenderTree.Add(componentId, frames);
-
-		public bool TryFindComponent<TComponent>(int parentComponentId, out int componentId, out TComponent component)
-		{
-			var result = false;
-			componentId = -1;
-			component = default!;
-
-			var frames = _currentRenderTree[parentComponentId];
-
-			for (var i = 0; i < frames.Count; i++)
-			{
-				ref var frame = ref frames.Array[i];
-				if (frame.FrameType == RenderTreeFrameType.Component)
-				{
-					if (frame.Component is TComponent c)
-					{
-						componentId = frame.ComponentId;
-						component = c;
-						result = true;
-						break;
-					}
-
-					if (TryFindComponent(frame.ComponentId, out componentId, out component))
-					{
-						result = true;
-						break;
-					}
-				}
-			}
-
-			return result;
-		}
 	}
 }
