+++ src/bunit.web/Rendering/RenderedComponent.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 
 using Microsoft.AspNetCore.Components;
+using Microsoft.AspNetCore.Components.RenderTree;
 
 namespace Bunit.Rendering
 {
@@ -22,9 +23,9 @@ namespace Bunit.Rendering
 			}
 		}
 
-		public RenderedComponent(int componentId, IServiceProvider services) : base(componentId, services) { }
+		internal RenderedComponent(int componentId, IServiceProvider services) : base(componentId, services) { }
 
-		public RenderedComponent(int componentId, TComponent instance, RenderTreeFrameCollection componentFrames, IServiceProvider services) : base(componentId, services)
+		internal RenderedComponent(int componentId, TComponent instance, RenderTreeFrameCollection componentFrames, IServiceProvider services) : base(componentId, services)
 		{
 			_instance = instance;
 			RenderCount++;
@@ -33,6 +34,8 @@ namespace Bunit.Rendering
 
 		protected override void OnRender(RenderEvent renderEvent)
 		{
+			// checks if this is the first render, and if it is
+			// tries to find the TCompoent in the render event			
 			if (_instance is null)
 			{
 				SetComponentAndID(renderEvent);
@@ -41,7 +44,7 @@ namespace Bunit.Rendering
 
 		private void SetComponentAndID(RenderEvent renderEvent)
 		{			
-			if (renderEvent.Frames.TryFindComponent<TComponent>(ComponentId, out var id, out var component))
+			if (TryFindComponent(renderEvent.Frames, ComponentId, out var id, out var component))
 			{
 				_instance = component;
 				ComponentId = id;
@@ -50,6 +53,38 @@ namespace Bunit.Rendering
 			{
 				throw new InvalidOperationException("Component instance not found at expected position in render tree.");
 			}
+		}
+
+		private bool TryFindComponent(RenderTreeFrameCollection framesCollection, int parentComponentId, out int componentId, out TComponent component)
+		{
+			var result = false;
+			componentId = -1;
+			component = default!;
+
+			var frames = framesCollection[parentComponentId];
+
+			for (var i = 0; i < frames.Count; i++)
+			{
+				ref var frame = ref frames.Array[i];
+				if (frame.FrameType == RenderTreeFrameType.Component)
+				{
+					if (frame.Component is TComponent c)
+					{
+						componentId = frame.ComponentId;
+						component = c;
+						result = true;
+						break;
+					}
+
+					if (TryFindComponent(framesCollection, frame.ComponentId, out componentId, out component))
+					{
+						result = true;
+						break;
+					}
+				}
+			}
+
+			return result;
 		}
 	}
 }
